third_party/blink/web_tests/platform/generic/fast/dom/HTMLAnchorElement/get-href-attribute-port-expected.txt

Test getting the port attribute of the URL in HTMLAnchorElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Default port is empty
PASS a.port is ''
Unspecified port should return empty string
PASS a.port is ''
Empty port should be empty
PASS a.port is ''
PASS successfullyParsed is true

TEST COMPLETE
